Cart and Login
Fixed remove button in cart.
Added mvc for logging in.
Fixed stock_id in CartDAO to use the right id.
Added a test user in the db.
Logging in does not currently function fully, but it is possible to log
in and out.
It is also possible to fully "checkout" which resets the cart.