Login & Register + Minor fixes
Implemented registering and logging in almost completely. Fixed some
minor issues.
To checkout the user must be logged in.
There is currently an issue with the session where if you restart the
server without clicking logout, when refreshing the page the user is
logged out without showing the login button.